Session control (#326)
Add controls for non-admin user sessions:

- Add jophiel_session_maxConcurrentSessionsPerUser config that prevent a user from logging in if the current session count for the user is greater than or equal to this limit.
- Add jophiel_session_disableLogout config that prevents a user from logging out if this config is true.
- Add session/batch-logout endpoint for admins to log out arbitrary users by JID.
- Display friendly error message for concurrent sessions exceeded error when logging in/logout disabled.

+++ judgels-backends/jophiel/jophiel-app/src/main/java/judgels/jophiel/session/SessionResource.java
@@ -1,13 +1,17 @@
 package judgels.jophiel.session;
 
+import static judgels.service.ServiceUtils.checkAllowed;
+
 import io.dropwizard.hibernate.UnitOfWork;
 import javax.inject.Inject;
 import javax.ws.rs.ForbiddenException;
+import judgels.jophiel.api.session.BatchLogoutData;
 import judgels.jophiel.api.session.Credentials;
 import judgels.jophiel.api.session.Session;
 import judgels.jophiel.api.session.SessionErrors;
 import judgels.jophiel.api.session.SessionService;
 import judgels.jophiel.api.user.User;
+import judgels.jophiel.user.UserRoleChecker;
 import judgels.jophiel.user.UserStore;
 import judgels.jophiel.user.account.UserRegistrationEmailStore;
 import judgels.service.actor.ActorChecker;
@@ -15,21 +19,27 @@ import judgels.service.api.actor.AuthHeader;
 
 public class SessionResource implements SessionService {
     private final ActorChecker actorChecker;
+    private final UserRoleChecker roleChecker;
     private UserStore userStore;
     private UserRegistrationEmailStore userRegistrationEmailStore;
     private SessionStore sessionStore;
+    private SessionConfiguration sessionConfiguration;
 
     @Inject
     public SessionResource(
             ActorChecker actorChecker,
+            UserRoleChecker roleChecker,
             UserStore userStore,
             UserRegistrationEmailStore userRegistrationEmailStore,
-            SessionStore sessionStore) {
+            SessionStore sessionStore,
+            SessionConfiguration sessionConfiguration) {
 
         this.actorChecker = actorChecker;
+        this.roleChecker = roleChecker;
         this.userStore = userStore;
         this.userRegistrationEmailStore = userRegistrationEmailStore;
         this.sessionStore = sessionStore;
+        this.sessionConfiguration = sessionConfiguration;
     }
 
     @Override
@@ -44,14 +54,39 @@ public class SessionResource implements SessionService {
             throw SessionErrors.userNotActivated(user.getEmail());
         }
 
+        if (!roleChecker.canAdminister(user.getJid())) {
+            int maxConcurrentSessionsPerUser = sessionConfiguration.getMaxConcurrentSessionsPerUser();
+            if (maxConcurrentSessionsPerUser >= 0) {
+                if (sessionStore.getSessionsByUserJid(user.getJid()).size() >= maxConcurrentSessionsPerUser) {
+                    throw SessionErrors.userMaxConcurrentSessionsExceeded(
+                            user.getUsername(), maxConcurrentSessionsPerUser);
+                }
+            }
+        }
+
         return sessionStore.createSession(SessionTokenGenerator.newToken(), user.getJid());
     }
 
     @Override
     @UnitOfWork
     public void logOut(AuthHeader authHeader) {
-        actorChecker.check(authHeader);
+        String actorJid = actorChecker.check(authHeader);
+        if (!roleChecker.canAdminister(actorJid) && sessionConfiguration.getDisableLogout()) {
+            throw SessionErrors.logoutDisabled();
+        }
+
         sessionStore.deleteSessionByToken(authHeader.getBearerToken());
         actorChecker.clear();
+    }
+
+    @Override
+    @UnitOfWork
+    public void batchLogout(AuthHeader authHeader, BatchLogoutData data) {
+        String actorJid = actorChecker.check(authHeader);
+        checkAllowed(roleChecker.canAdminister(actorJid));
+
+        for (String userJid : data.getUserJids()) {
+            sessionStore.deleteSessionsByUserJid(userJid);
+        }
     }
 }
